Event List & Event Details Pages
Event Details page contains all the controls as well as empty hyper
links in case we want to do something fancy with them and show events
scheduled for a certain day, or building, or class, etc. Also a blank
space for a map or picture or some content.
Event List page has a table that can be easily populated with data and
the formatting will be fine.
Also included in this was an update to the navigation and controller as
needed.
New favicons were also in here.
Lastly, the testing page was moved into the new Events navigation items
submenu and it's references were renamed.